Renamed util module to geom.
